Add jansi and jline jars to the top loader
sbt needs to use a forked version of jline to ensure that the scala
console works with the remote client. Since jline and jansi are both
java only jars, they can safely be put in our top loader so that all
descendant classloaders will pick up the forked behavior. By
constructing the classloaders this way in the launcher, we can short
circuit sbt main from creating a new classloader hierarchy to ensure
this behavior. It will check that there is a getEarlyJars method defined
on the top classloader using reflection and then verify that the three
jars that it finds from that method are what it expects them to be.