Do not create jansi classloader in windows
For historical reasons, it was necessary to always put jansi in a
classloader above the rest of the metabuild classpath. As of sbt 1.x,
that is no longer the case and causes problems when using the latest
jansi version in sbt. We still need to have the jansi loader for legacy
versions so now we only make the jansi loader if we detect that we are
launching a 0.x version on non cygwin windows.