Workaround NullPointerException issue
Problem
-------
We started getting reports of NPE because the boot directory contains
test-interface.jar as opposed to test-interface-1.0.jar.
I am not fully sure what the cause is, but maybe Coursier has changed
its cache structure over the years?

Solution
--------
A workaround for now would be to guess the version number from the
directory name, if not call it test-interface-0.0.jar.