Override Scala version for sbt 1.4.x
Problem
-------
During sbt 1.4.x series util-interface incorrectly depended on
scala-library. Specifically for util-interface 1.4.0, 1.4.1, and 1.4.2,
this was Scala 2.13. Ivy seemed to be ok with this, but Coursier evicts
Scala 2.12 and resolves to 2.13, which creating a confusing
`java.lang.NoClassDefFoundError: scala/Serializable` error.

Solution
--------
Explicitly override Scala version for sbt 1.4.x to Scala 2.12.12.